Capitalise first letter of enum names in defaults
When defining the enums themselves, the first letter is capitalised in
order to create a constant named after the enum, however if a message
containing that enum has a default value set, the first letter of *that*
wasn't being capitalised, causing an error.  This commit ensures that
both are capitalised.